hom view 리펙토링
