+++ core/src/main/java/com/example/core/TopAppBarComponent.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
@@ -73,6 +74,7 @@ fun TopAppBarWithCloseButton(
     CenterAlignedTopAppBar(
         modifier = Modifier
             .fillMaxWidth()
+            .statusBarsPadding()
             .border(width = 1.dp, color = ColorStyle.GRAY_200),
         title = {
             Text(
