+++ presenter/main/src/main/java/com/sungil/main/MainActivity.kt
@@ -4,12 +4,9 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.runtime.getValue
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.sungil.domain.model.Router
 import com.sungil.main.ui.MainScreenView
-import com.sungil.main.ui.review.ReviewView
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
@@ -23,35 +20,24 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             val navController = rememberNavController()
-            val navBackStackEntry by navController.currentBackStackEntryAsState()
-            val currentRoute = navBackStackEntry?.destination?.route
-
-            if (currentRoute == MainView.REVIEW.route) {
-                ReviewView(
-                    onClose = { navController.popBackStack() },
-                    viewModel = viewModel
-                )
-            } else {
-                MainScreenView(
-                    navController = navController,
-                    viewModel = viewModel,
-                    profileButtonClick = { router.navigation(NAV_EDIT_PROFILE) },
-                    reportClick = { router.navigation(NAV_REPORT) },
-                    lowClick = { router.navigation(NAV_LOW) },
-                    alarmClick = { router.navigation(NAV_ALARM) },
-                    oneThingClick = { router.navigation(NAV_ONE_THING) },
-                    firstMatchClick = { destination ->
-                        val bundle = Bundle().apply {
-                            putString(BuildConfig.KEY_MATCH, destination)
-                        }
-                        router.navigation(NAV_FIRST_MATCH, bundle)
-                    },
-                    randomMatchClick = { router.navigation(NAV_RANDOM_MATCH) },
-                    login = { router.navigation(NAV_LOGIN) },
-                    guide = { router.navigation(NAV_GUIDE) }
-                )
-            }
+            MainScreenView(
+                navController = navController,
+                viewModel = viewModel,
+                profileButtonClick = { router.navigation(NAV_EDIT_PROFILE) },
+                reportClick = { router.navigation(NAV_REPORT) },
+                lowClick = { router.navigation(NAV_LOW) },
+                alarmClick = { router.navigation(NAV_ALARM) },
+                oneThingClick = { router.navigation(NAV_ONE_THING) },
+                firstMatchClick = { destination ->
+                    val bundle = Bundle().apply {
+                        putString(BuildConfig.KEY_MATCH, destination)
+                    }
+                    router.navigation(NAV_FIRST_MATCH, bundle)
+                },
+                randomMatchClick = { router.navigation(NAV_RANDOM_MATCH) },
+                login = { router.navigation(NAV_LOGIN) },
+                guide = { router.navigation(NAV_GUIDE) }
+            )
         }
-
     }
 }
