+++ presenter/main/src/main/java/com/sungil/main/nav/MainNavigation.kt
@@ -1,14 +1,18 @@
 package com.sungil.main.nav
 
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
 import com.sungil.main.MainViewModel
 import com.sungil.main.BottomView
 import com.sungil.main.MainView
 import com.sungil.main.ui.myMatch.MyMatchView
-import com.sungil.main.ui.home.HomeScreen
+import com.sungil.main.ui.home.HomeViewScreen
 import com.sungil.main.ui.matchDetail.MeetDetailView
 import com.sungil.main.ui.myPage.MyPageScreen
 import com.sungil.main.ui.payDetail.PayDetailView
@@ -27,19 +31,35 @@ fun MainNavigation(
     randomMatchClick: () -> Unit,
     login: () -> Unit,
     guide: () -> Unit,
+    paddingValues: PaddingValues,
+    snackBarHostState: SnackbarHostState,
 ) {
-    NavHost(navController = navController, startDestination = BottomView.Home.screenRoute) {
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentRoute = navBackStackEntry?.destination?.route
 
+    NavHost(navController = navController, startDestination = BottomView.Home.screenRoute) {
         composable(BottomView.Home.screenRoute) {
-            HomeScreen(
+            HomeViewScreen(
                 viewModel = viewModel,
-                alarmClick = alarmClick,
-                notifyClick = {},
-                oneThingMatchClick = oneThingClick,
-                randomMatchClick = randomMatchClick,
                 firstMatchClick = firstMatchClick,
-                login = login
+                notifyClick = {
+                    //TODO 출시 직전 구현 완료
+                },
+                oneThingClick = oneThingClick,
+                paddingValues = paddingValues,
+                randomMatchClick = randomMatchClick,
+                reLogin = login,
+                snackBarHostState = snackBarHostState
             )
+//            HomeViewScreen(
+//                viewModel = viewModel,
+//                alarmClick = alarmClick,
+//                notifyClick = {},
+//                oneThingMatchClick = oneThingClick,
+//                randomMatchClick = randomMatchClick,
+//                firstMatchClick = firstMatchClick,
+//                login = login
+//            )
         }
 
         composable(BottomView.Calendar.screenRoute) {
@@ -75,13 +95,12 @@ fun MainNavigation(
                 viewModel = viewModel
             )
         }
-
         composable(MainView.REVIEW.route) {
             ReviewView(
                 onClose = { navController.popBackStack() },
-                viewModel = viewModel
+                viewModel = viewModel,
+                paddingValues = paddingValues
             )
         }
-
     }
 }
