--- presenter/main/src/main/res/drawable/ic_light_yellow.xml
@@ -1,22 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="28dp"
-    android:height="27dp"
-    android:viewportWidth="28"
-    android:viewportHeight="27">
-  <group>
-    <clip-path
-        android:pathData="M0,0.5h28v26h-28z"/>
-    <path
-        android:pathData="M19.562,9.809H13.271V2.414C13.271,0.609 11.161,-0.183 10.106,1.211L0.395,14.088C-0.545,15.332 0.281,17.205 1.781,17.205H8.072V24.6C8.072,26.404 10.182,27.197 11.237,25.803L20.935,12.94C21.875,11.696 21.049,9.823 19.549,9.823L19.562,9.809Z"
-        android:fillColor="#FFCF6A"/>
-    <path
-        android:pathData="M27.225,2.195H18.812C18.376,2.195 18.023,2.575 18.023,3.043V4.505C18.023,4.973 18.376,5.353 18.812,5.353H27.225C27.661,5.353 28.014,4.973 28.014,4.505V3.043C28.014,2.575 27.661,2.195 27.225,2.195Z"
-        android:fillColor="#FF8E62"/>
-    <path
-        android:pathData="M27.225,21.524H18.812C18.376,21.524 18.023,21.904 18.023,22.372V23.834C18.023,24.302 18.376,24.682 18.812,24.682H27.225C27.661,24.682 28.014,24.302 28.014,23.834V22.372C28.014,21.904 27.661,21.524 27.225,21.524Z"
-        android:fillColor="#FF8E62"/>
-    <path
-        android:pathData="M27.226,11.86H23.808C23.372,11.86 23.02,12.239 23.02,12.707V14.17C23.02,14.638 23.372,15.017 23.808,15.017H27.226C27.662,15.017 28.014,14.638 28.014,14.17V12.707C28.014,12.239 27.662,11.86 27.226,11.86Z"
-        android:fillColor="#FF8E62"/>
-  </group>
-</vector>
