+++ presenter/main/src/main/res/drawable/ic_lighting_match.xml
@@ -1,0 +1,25 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="37dp"
+    android:viewportWidth="40"
+    android:viewportHeight="37">
+  <group>
+    <clip-path
+        android:pathData="M0,0.5h40v36h-40z"/>
+    <path
+        android:pathData="M38.946,12.638H24.664C24.078,12.638 23.609,12.166 23.609,11.592V10.059C23.609,9.478 24.085,9.013 24.664,9.013H38.946C39.532,9.013 40.001,9.486 40.001,10.059V11.592C40.001,12.166 39.532,12.638 38.946,12.638Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M38.946,20.315H29.461C28.875,20.315 28.406,19.843 28.406,19.27V17.737C28.406,17.156 28.882,16.691 29.461,16.691H38.946C39.532,16.691 40,17.163 40,17.737V19.27C40,19.843 39.532,20.315 38.946,20.315Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M38.946,27.993H24.664C24.078,27.993 23.609,27.521 23.609,26.947V25.414C23.609,24.833 24.085,24.368 24.664,24.368H38.946C39.532,24.368 40.001,24.84 40.001,25.414V26.947C40.001,27.528 39.532,27.993 38.946,27.993Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M16.848,0.798L0.215,18.31C-0.275,18.87 0.127,19.741 0.874,19.741H17.507C17.991,19.741 18.379,19.356 18.379,18.877V1.365C18.379,0.566 17.376,0.195 16.848,0.798Z"
+        android:fillColor="#FFCF6A"/>
+    <path
+        android:pathData="M8.742,36.201L25.375,18.688C25.865,18.129 25.463,17.258 24.716,17.258H8.083C7.599,17.258 7.211,17.642 7.211,18.122V35.634C7.211,36.433 8.214,36.804 8.742,36.201Z"
+        android:fillColor="#FFCF6A"/>
+  </group>
+</vector>
