+++ presenter/main/src/main/res/drawable/ic_one_thing_match.xml
@@ -1,0 +1,40 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="40dp"
+    android:viewportWidth="40"
+    android:viewportHeight="40">
+  <group>
+    <clip-path
+        android:pathData="M0,0h40v40h-40z"/>
+    <path
+        android:pathData="M10.932,40C16.969,40 21.863,35.273 21.863,29.443C21.863,23.612 16.969,18.885 10.932,18.885C4.894,18.885 0,23.612 0,29.443C0,35.273 4.894,40 10.932,40Z"
+        android:fillColor="#FFCF6A"/>
+    <path
+        android:pathData="M11.129,31.512H2.445V40H11.129V31.512Z"
+        android:fillColor="#FFCF6A"/>
+    <path
+        android:pathData="M13.973,29.443C13.973,30.244 14.606,30.892 15.451,30.892C16.296,30.892 16.93,30.244 16.93,29.443C16.93,28.641 16.296,27.993 15.451,27.993C14.606,27.993 13.973,28.641 13.973,29.443Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M9.453,29.443C9.453,30.244 10.087,30.892 10.932,30.892C11.777,30.892 12.41,30.244 12.41,29.443C12.41,28.641 11.777,27.993 10.932,27.993C10.087,27.993 9.453,28.641 9.453,29.443Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M4.941,29.443C4.941,30.244 5.575,30.892 6.42,30.892C7.265,30.892 7.899,30.244 7.899,29.443C7.899,28.641 7.265,27.993 6.42,27.993C5.575,27.993 4.941,28.641 4.941,29.443Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M36.536,16.926H24.902V28.299H36.536V16.926Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M25.346,28.299C33.438,28.299 39.998,21.964 39.998,14.149C39.998,6.335 33.438,0 25.346,0C17.255,0 10.695,6.335 10.695,14.149C10.695,21.964 17.255,28.299 25.346,28.299Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M29.016,14.149C29.016,15.082 29.755,15.832 30.736,15.832C31.716,15.832 32.456,15.075 32.456,14.149C32.456,13.224 31.716,12.466 30.736,12.466C29.755,12.466 29.016,13.217 29.016,14.149Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M23.629,14.149C23.629,15.082 24.368,15.832 25.349,15.832C26.33,15.832 27.069,15.075 27.069,14.149C27.069,13.224 26.33,12.466 25.349,12.466C24.368,12.466 23.629,13.217 23.629,14.149Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M18.242,14.149C18.242,15.082 18.982,15.832 19.962,15.832C20.943,15.832 21.682,15.075 21.682,14.149C21.682,13.224 20.943,12.466 19.962,12.466C18.982,12.466 18.242,13.217 18.242,14.149Z"
+        android:fillColor="#61C9A0"/>
+  </group>
+</vector>
