--- presenter/main/src/main/res/drawable/ic_one_thing_purple.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="28dp"
-    android:height="28dp"
-    android:viewportWidth="28"
-    android:viewportHeight="28">
-  <group>
-    <clip-path
-        android:pathData="M0,0.5h28v27h-28z"/>
-    <path
-        android:pathData="M28,11.856C28,5.583 21.734,0.5 14,0.5C6.266,0.5 0,5.583 0,11.856C0,18.129 6.266,23.211 14,23.211V26.798C14,27.271 14.458,27.622 14.917,27.485C28.397,23.425 28,11.856 28,11.856ZM6.771,14.191C5.472,14.191 4.417,13.138 4.417,11.84C4.417,10.543 5.472,9.49 6.771,9.49C8.07,9.49 9.124,10.543 9.124,11.84C9.124,13.138 8.07,14.191 6.771,14.191ZM14,14.191C12.701,14.191 11.646,13.138 11.646,11.84C11.646,10.543 12.701,9.49 14,9.49C15.299,9.49 16.354,10.543 16.354,11.84C16.354,13.138 15.299,14.191 14,14.191ZM21.229,14.191C19.93,14.191 18.875,13.138 18.875,11.84C18.875,10.543 19.93,9.49 21.229,9.49C22.528,9.49 23.583,10.543 23.583,11.84C23.583,13.138 22.528,14.191 21.229,14.191Z"
-        android:fillColor="#6700CE"/>
-  </group>
-</vector>
