--- presenter/main/src/main/res/drawable/ic_random_green.xml
@@ -1,16 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="32dp"
-    android:height="29dp"
-    android:viewportWidth="32"
-    android:viewportHeight="29">
-  <group>
-    <clip-path
-        android:pathData="M0,0h32v29h-32z"/>
-    <path
-        android:pathData="M29.338,10.527H24.81C24.555,10.527 24.362,10.793 24.446,11.046C24.761,12.007 24.555,13.121 23.817,13.893L14.773,23.344C14.482,23.648 14.313,24.066 14.313,24.496V26.242C14.313,27.773 15.499,29.025 16.976,29.025H29.338C30.803,29.025 32.001,27.785 32.001,26.242V13.323C32.001,11.792 30.815,10.54 29.338,10.54V10.527ZM23.163,21.13C22.449,21.13 21.855,20.523 21.855,19.764C21.855,19.004 22.437,18.397 23.163,18.397C23.889,18.397 24.471,19.004 24.471,19.764C24.471,20.523 23.889,21.13 23.163,21.13Z"
-        android:fillColor="#109F67"/>
-    <path
-        android:pathData="M22.026,9.958L13.284,0.81C12.243,-0.278 10.56,-0.278 9.519,0.81L0.777,9.958C-0.264,11.046 -0.264,12.805 0.777,13.893L9.519,23.028C10.56,24.117 12.243,24.117 13.284,23.028L22.026,13.893C23.067,12.805 23.067,11.046 22.026,9.958ZM5.838,12.856C5.329,13.387 4.506,13.387 3.998,12.856C3.489,12.324 3.489,11.464 3.998,10.932C4.506,10.401 5.329,10.401 5.838,10.932C6.346,11.464 6.346,12.324 5.838,12.856ZM10.487,4.151C10.996,3.619 11.819,3.619 12.328,4.151C12.836,4.682 12.836,5.542 12.328,6.074C11.819,6.605 10.996,6.605 10.487,6.074C9.979,5.542 9.979,4.682 10.487,4.151ZM10.487,10.958C10.996,10.426 11.819,10.426 12.328,10.958C12.836,11.489 12.836,12.349 12.328,12.881C11.819,13.412 10.996,13.412 10.487,12.881C9.979,12.349 9.979,11.489 10.487,10.958ZM12.352,19.663C11.843,20.194 11.02,20.194 10.511,19.663C10.003,19.131 10.003,18.271 10.511,17.74C11.02,17.208 11.843,17.208 12.352,17.74C12.86,18.271 12.86,19.131 12.352,19.663ZM18.841,12.881C18.333,13.412 17.51,13.412 17.001,12.881C16.493,12.349 16.493,11.489 17.001,10.958C17.51,10.426 18.333,10.426 18.841,10.958C19.35,11.489 19.35,12.349 18.841,12.881Z"
-        android:fillColor="#61C9A0"/>
-  </group>
-</vector>
