+++ presenter/main/src/main/res/drawable/ic_random_match.xml
@@ -1,0 +1,34 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="41dp"
+    android:viewportWidth="40"
+    android:viewportHeight="41">
+  <group>
+    <clip-path
+        android:pathData="M0,0.5h40v40h-40z"/>
+    <path
+        android:pathData="M16.858,38.716L16.758,19.018C16.751,18.035 17.548,17.219 18.531,17.219L38.112,17.118C39.089,17.111 39.9,17.913 39.9,18.902L40.001,38.6C40.008,39.583 39.211,40.399 38.227,40.399L18.646,40.5C17.67,40.5 16.866,39.698 16.858,38.716Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M27.139,27.57C26.45,28.271 26.428,29.369 27.132,30.069C27.836,30.77 28.927,30.734 29.617,30.033C30.306,29.333 30.327,28.235 29.624,27.534C28.92,26.833 27.829,26.87 27.139,27.57Z"
+        android:fillColor="#5D1CC4"/>
+    <path
+        android:pathData="M14.251,31.659L0.507,17.407C-0.182,16.692 -0.168,15.536 0.543,14.843L14.717,1.01C15.428,0.317 16.577,0.331 17.266,1.046L31.01,15.305C31.699,16.02 31.685,17.176 30.974,17.869L16.807,31.702C16.096,32.396 14.947,32.381 14.251,31.659Z"
+        android:fillColor="#6700CE"/>
+    <path
+        android:pathData="M17.431,9.736C17.445,8.833 16.749,8.14 15.887,8.132C15.018,8.118 14.307,8.79 14.293,9.693C14.279,10.595 14.975,11.289 15.837,11.296C16.706,11.303 17.417,10.639 17.431,9.736Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M17.232,23.027C17.246,22.124 16.55,21.43 15.688,21.423C14.819,21.409 14.108,22.08 14.094,22.983C14.08,23.886 14.776,24.58 15.638,24.587C16.507,24.594 17.218,23.93 17.232,23.027Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M22.344,18.035C23.242,18.05 23.931,17.349 23.938,16.483C23.952,15.608 23.285,14.893 22.387,14.879C21.49,14.864 20.8,15.565 20.793,16.432C20.786,17.299 21.446,18.021 22.344,18.035Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M9.129,17.833C10.027,17.847 10.716,17.147 10.723,16.28C10.738,15.406 10.07,14.691 9.172,14.677C8.275,14.662 7.585,15.363 7.578,16.229C7.571,17.104 8.239,17.819 9.129,17.833Z"
+        android:fillColor="#61C9A0"/>
+    <path
+        android:pathData="M15.738,17.934C16.636,17.949 17.325,17.248 17.333,16.381C17.347,15.507 16.679,14.792 15.782,14.778C14.884,14.763 14.195,15.464 14.188,16.331C14.18,17.197 14.841,17.92 15.738,17.934Z"
+        android:fillColor="#61C9A0"/>
+  </group>
+</vector>
